fix warnings

+++ rmw/include/rmw/types.h
@@ -20,6 +20,7 @@ extern "C"
 {
 #endif
 
+#include <stddef.h>
 #include <stdint.h>
 
 #include "visibility_control.h"
@@ -66,25 +67,25 @@ typedef struct RMW_PUBLIC_TYPE rmw_guard_condition_t
 
 typedef struct RMW_PUBLIC_TYPE rmw_subscriptions_t
 {
-  unsigned long subscriber_count;
+  size_t subscriber_count;
   void ** subscribers;
 } rmw_subscriptions_t;
 
 typedef struct RMW_PUBLIC_TYPE rmw_services_t
 {
-  unsigned long service_count;
+  size_t service_count;
   void ** services;
 } rmw_services_t;
 
 typedef struct RMW_PUBLIC_TYPE rmw_clients_t
 {
-  unsigned long client_count;
+  size_t client_count;
   void ** clients;
 } rmw_clients_t;
 
 typedef struct RMW_PUBLIC_TYPE rmw_guard_conditions_t
 {
-  unsigned long guard_condition_count;
+  size_t guard_condition_count;
   void ** guard_conditions;
 } rmw_guard_conditions_t;
